add settings panel in pause and checkbox in settings
