+++ Assets/Fonts/TMPfont/jelani SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8d47a53845212314694450c94884fbd0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
