+++ Assets/Prefabs/Player/Animation/MonkeyPlayer.controller
@@ -66,9 +66,9 @@ AnimatorStateTransition:
   m_Mute: 0
   m_IsExit: 0
   serializedVersion: 3
-  m_TransitionDuration: 0.0996537
-  m_TransitionOffset: 0
-  m_ExitTime: 0.000089394365
+  m_TransitionDuration: 0
+  m_TransitionOffset: 0.8959022
+  m_ExitTime: 0.00008939446
   m_HasExitTime: 0
   m_HasFixedDuration: 1
   m_InterruptionSource: 0
@@ -91,10 +91,10 @@ AnimatorStateTransition:
   m_Mute: 0
   m_IsExit: 0
   serializedVersion: 3
-  m_TransitionDuration: 0.6989565
+  m_TransitionDuration: 0.12446792
   m_TransitionOffset: 0
-  m_ExitTime: 0.000000034059795
-  m_HasExitTime: 1
+  m_ExitTime: 0.000000049851867
+  m_HasExitTime: 0
   m_HasFixedDuration: 1
   m_InterruptionSource: 0
   m_OrderedInterruption: 1
