+++ Assets/Scripts/PausePanel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 262a2969ccd824a40b6d396567d6dc60
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
