+++ Assets/Scripts/SettingsPanel.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 398490ae22311494186f6a623c16e016
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
