+++ Assets/Scripts/SettingsPanel/SettingsPanel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6d1c59a1a63d9db4ea383729bb1805a0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
